chore: now seeding e2e user in normal seed script
